requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.3
Brotli
certifi
charset-normalizer
click==8.1.7
colorama==0.4.6
Cython==3.0.11
docopt==0.6.2
filelock
flexcache==0.3
flexparser==0.3.1
fsspec==2024.9.0
gmpy2
greenlet==3.1.1
gTTS==2.5.3
huggingface-hub
idna==3.10
Jinja2
joblib==1.4.2
Js2Py==0.74
MarkupSafe
mathparse==0.1.2
mkl-service==2.4.0
mkl_fft
mkl_random
mpmath
networkx
nltk==3.9.1
numpy
packaging==24.1
pillow
Pint==0.24.3
pipwin==0.5.2
playsound==1.3.0
PyAudio==0.2.14
pydub==0.25.1
pyjsparser==2.7.1
pymongo==3.13.0
PyPrind==2.11.3
pySmartDL==1.3.4
PySocks
python-dateutil==2.7.5
python-dotenv==1.0.1
pytz==2024.2
PyYAML
regex==2024.9.11
requests
safetensors==0.4.5
six==1.16.0
soupsieve==2.6
SpeechRecognition==3.10.4
SQLAlchemy==1.2.5
sympy
tokenizers==0.20.1
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers
typing_extensions
tzdata==2024.2
tzlocal==5.2
urllib3
win-inet-pton
